move firebase to plugins
