+++ hosting/src/plugins/firebase.ts
